feat(editor): export `getContainerChildren` from `/traversal`
`plugin-list-index`, and any consumer that descends editable containers,
needs to walk a node's children without re-deriving the path and
re-walking from the root at each level, which is O(depth^2) for
recursive descent. The public `getChildren` is path-based and re-walks.
The node-based resolver was only available as `getNodeChildren`, the
internal descent kernel the positional traversal utilities share, which
also folds the text-block and root-document cases and so carries a
`Node | {value}` parameter union no container consumer wants.

Extract the container arm of that kernel as `getContainerChildren`
(`@beta`) and export it from `@portabletext/editor/traversal`. Given a
node and optionally its resolved parent container, it returns
`{children, container}` for a registered container and `undefined` for
anything else (text blocks, spans, leaves, unregistered objects),
resolving in O(1) per level. `children` is the editable child array;
`container` is the node's own container registration, off which the
caller reads `container.field.name` (the path segment) and which it
threads back as the parent for the next level. `getNodeChildren` stays
internal and now delegates its container case here, so the resolution
logic lives in one place. No behavior change.
