+++ packages/editor/src/traversal/index.ts
@@ -2,6 +2,7 @@ export {getAncestor} from './get-ancestor'
 export {getAncestors} from './get-ancestors'
 export {getBlock, isBlock} from './is-block'
 export {getChildren} from './get-children'
+export {getContainerChildren} from './get-container-children'
 export {getEnclosingBlock} from './get-enclosing-block'
 export {getFirstChild} from './get-first-child'
 export {getLastChild} from './get-last-child'
